refactor: reorganize imports and remove unneeded files
- se reorganizaron los requerimientos
- se ajustaron las importaciones
- se configuraron los modulos del proyecto
- se configuraron SonarQube y pytest
