+++ src/controllers/PaymentsManager.py
@@ -1,4 +1,4 @@
-from .Manager import Manager
+from controllers import Manager
 
 
 class PaymentsManager(Manager):
