+++ src/controllers/__init__.py
@@ -2,10 +2,10 @@ from src.controllers.ShipmentManager import ShipmentManager
 from src.controllers.PackagesManager import PackageManager
 from src.controllers.LoginManager import LoginManager
 from src.controllers.PaymentsManager import PaymentsManager
-from src.controllers.UsersManager import UsersManager
+from src.controllers.UsersManager import UserManager
 
 __all__ = ["ShipmentManager",
            "PackageManager",
            "LoginManager",
            "PaymentsManager",
-           "UsersManager"]
+           "UserManager"]
