+++ src/utils/__init__.py
@@ -1,0 +1,9 @@
+from utils.utils import validate_input, generate_id
+
+__all__ = ["validate_input", "generate_id"]
+
+NO_FOUND = "registro no encontrado"
+LOGIN_OPTION = "1"
+REGISTER_OPTION = "2"
+EXIT_OPTION = "q"
+MESSAGGE_MAIN_MENU = "Regresando al menu principal..."
