+++ tests/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+pythonpath = C:\Users\ACER\Documents\Proyectos - Software\Python\2025\SGPaquetes\src
+
